feat(build_path): is_executable_commandをis_directoryとis_executableに分割
それぞれsymbolic linkを参照しなかったためlstatでなくstatを使用。
build_path.cの47行目は25行に収まらなかったため{}を使わずに書いています。
